try all the possible replacements
a type variable may be equivalent to multiple expressions, but we were
arbitrarily picking a single one to replace it with. now we're trying
all the possibilities.
